+++ nextjs/components/common/Footer.tsx
