+++ nextjs/components/common/Hero.tsx
