+++ nextjs/components/common/VideoPlayer.tsx
