--- nextjs/pages/404.ts
@@ -1,3 +1,0 @@
-import getErrorPage from "../components/pages/getErrorPage";
-
-export default getErrorPage(404);
