+++ nextjs/pages/404.tsx
@@ -1,0 +1,11 @@
+import Page from "../components/Page";
+
+export default function ErrorPage() {
+  return (
+    <Page>
+      <div className="my-12 text-center text-4xl text-yellow-600">
+        Oops! You have found a missing page.
+      </div>
+    </Page>
+  );
+}
