--- nextjs/pages/index.js
@@ -1,3 +1,0 @@
-import IndexPage from "../components/pages";
-
-export default IndexPage;
