--- nextjs/pages/news.js
@@ -1,1 +1,0 @@
-export {default, getStaticProps} from "../components/pages/news";
