--- nextjs/pages/news/[id].ts
@@ -1,5 +1,0 @@
-export {
-  default,
-  getStaticProps,
-  getStaticPaths,
-} from "../../components/pages/newsItem";
