--- nextjs/pages/opportunities.ts
@@ -1,1 +1,0 @@
-export { default } from "../components/pages/opportunities";
