--- nextjs/pages/opportunities/[id].ts
@@ -1,1 +1,0 @@
-export { default } from "../../components/pages/opportunityItem";
