--- nextjs/pages/talks.js
@@ -1,1 +1,0 @@
-export { default } from "../components/pages/talks";
