--- nextjs/pages/talks/[id].ts
@@ -1,1 +1,0 @@
-export { default } from "../../components/pages/talkItem";
